Test fetching state
I dropped the associated enum error as it isn’t being used, and for testing is just adding complexity.
In fact, we could have an associated value for the failed state with a String which could be the message to display.
Currently we are not doing that so I’ll not add any.

Signed-off-by: Rui Barbosa <rui.barbosa@hey.com>